refactor: Relocate event methods: OnDamaged & OnDestroy to the SAlien base class, instead of individual.
