+++ src/StardustDefender/Entities/Aliens/SAlien.cs
@@ -1,11 +1,42 @@
 ﻿using Microsoft.Xna.Framework;
 
 using StardustDefender.Controllers;
+using StardustDefender.Effects.Common;
+using StardustDefender.Engine;
+using StardustDefender.Managers;
+
+using System.Threading.Tasks;
 
 namespace StardustDefender.Entities.Aliens
 {
     internal abstract class SAlien : SEntity
     {
+        protected override void OnDamaged(int value)
+        {
+            _ = SSounds.Play("Damage_02");
+            _ = SEffectsManager.Create<SImpactEffect>(WorldPosition);
+
+            _ = Task.Run(async () =>
+            {
+                Color = Color.Red;
+                await Task.Delay(235);
+                Color = Color.White;
+            });
+        }
+        protected override void OnDestroy()
+        {
+            SLevelController.EnemyKilled();
+
+            _ = SSounds.Play("Explosion_01");
+            _ = SEffectsManager.Create<SExplosionEffect>(WorldPosition);
+
+            // Drop
+            if (SRandom.Chance(20, 100))
+            {
+                _ = SItemsManager.GetRandomItem(WorldPosition);
+            }
+        }
+
         protected void CollideWithPlayer()
         {
             if (Vector2.Distance(SLevelController.Player.WorldPosition, WorldPosition) < 32)
