+++ src/StardustDefender/StardustDefender.WindowsDX.csproj
@@ -25,6 +25,9 @@
 		<PackageReference Include="MonoGame.Framework.WindowsDX" Version="3.8.1.303" />
 		<PackageReference Include="MonoGame.Content.Builder.Task" Version="3.8.1.303" />
 	</ItemGroup>
+	<ItemGroup>
+	  <Folder Include="Entities\Attributes\" />
+	</ItemGroup>
 
 	<!-- TARGETS -->
 	<Target Name="RestoreDotnetTools" BeforeTargets="Restore">
